+++ packages/storybook/examples/core/Icon/BasicIcons.js
@@ -1,43 +1,19 @@
 import React from 'react';
-
 import Icon from '@ichef/gypcrete/src/Icon';
+
+import iconSet from '../../../../../configs/fontello.config.json';
 
 function BasicIconsSet() {
     return (
-        <div>
-            <Icon type="add" />
-            <Icon type="edit" />
-            <Icon type="duplicate" />
-            <Icon type="copypaste" />
-            <Icon type="tickets" />
-            <Icon type="trashcan" />
-            <Icon type="drag" />
-            <Icon type="success" />
-            <Icon type="error" />
-            <Icon type="delete" />
-            <Icon type="header-next" />
-            <Icon type="header-prev" />
-            <Icon type="fold" />
-            <Icon type="unfold" />
-            <Icon type="next" />
-            <Icon type="prev" />
-            <Icon type="folder" />
-            <Icon type="loading" spinning />
-            <Icon type="lock" />
-            <Icon type="unlock" />
-            <Icon type="radio-empty" />
-            <Icon type="radio-half" />
-            <Icon type="radio-selected" />
-            <Icon type="picture" />
-            <Icon type="printer" />
-            <Icon type="pause" />
-            <Icon type="play" />
-            <Icon type="search" />
-            <Icon type="download" />
-            <Icon type="dropdown" />
-            <Icon type="remove-element" />
-            <Icon type="inline-info" />
-        </div>
+        <>
+            {
+                iconSet.glyphs.map(({ css: iconName }) => (
+                    <Icon
+                        type={iconName}
+                        spinning={iconName.includes('loading')} />
+                ))
+            }
+        </>
     );
 }