+++ CHANGELOG.md
@@ -11,9 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - [Storybook] Add and configure storybook docs addon.
+- [Core] Add new icon `mapping`. (#241)
 
 ### Changed
 - [Storybook] Revamp docs for `form` and `imageeditor` packages.
+- [Storybook] Modify icon set source of `Basic icons set` page from `configs/fontello.config.json`. (#241)
+- [Storybook] Fix Icon props doc page. (#241)
 
 
 ## [4.0.0]